Move WalkerAlias to utils
